# bff-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for Frontend (BFF) service for BankEase mobile banking application. Single entry point for the mobile app, orchestrating calls to identity-service and user-profile-service.",
        "title": "BFF Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/currency-list": {
            "get": {
                "description": "Retrieve all supported currency exchange entries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Get currency list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.CurrencyListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            }
        },
...
        "/api/mobile-prepaid/beneficiaries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve saved mobile prepaid top-up contacts for an account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile Prepaid"
                ],
                "summary": "Get prepaid beneficiaries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.BeneficiaryListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Save a new mobile prepaid top-up contact",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile Prepaid"
                ],
                "summary": "Add beneficiary",
                "parameters": [
                    {
                        "description": "Add beneficiary request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.AddBeneficiaryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.BeneficiaryItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            }
        },
        "/api/mobile-prepaid/beneficiaries/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search saved mobile prepaid contacts by name or phone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile Prepaid"
                ],
                "summary": "Search beneficiaries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.BeneficiaryListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            }
        },
        "/api/mobile-prepaid/cards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all payment cards for an account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile Prepaid"
                ],
                "summary": "Get payment cards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.PaymentCardListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new payment card/account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile Prepaid"
                ],
                "summary": "Create payment card",
                "parameters": [
                    {
                        "description": "Create card request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.CreatePaymentCardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.PaymentCardItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            }
        },
        "/api/mobile-prepaid/pay": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Process a mobile prepaid top-up transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile Prepaid"
                ],
                "summary": "Submit prepaid payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client-generated UUID for idempotency",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Prepaid payment request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.PrepaidPayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.PrepaidPayResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            }
        },
        "/api/pay-the-bill/internet-bill": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the authenticated user's internet bill detail. Requires JWT authentication.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Get internet bill detail",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.InternetBillResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            }
        },
        "/api/pay-the-bill/providers": {
            "get": {
                "description": "Retrieve all internet service providers available for bill payment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Get internet bill providers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ProviderListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_bankease_bff-service_protogen_bff-service.AddBeneficiaryRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.BeneficiaryItem": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.BeneficiaryListResponse": {
            "type": "object",
            "properties": {
                "beneficiaries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.BeneficiaryItem"
                    }
                }
            }
        },
...
        "github_com_bankease_bff-service_protogen_bff-service.CreatePaymentCardRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "balance": {
                    "type": "integer"
                },
                "brand": {
                    "type": "string"
                },
                "cardLabel": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "gradientColors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "holderName": {
                    "type": "string"
                },
                "maskedNumber": {
                    "type": "string"
                }
            }
        },
...
        "github_com_bankease_bff-service_protogen_bff-service.CurrencyEntry": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.CurrencyListResponse": {
            "type": "object",
            "properties": {
                "currencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.CurrencyEntry"
                    }
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.ErrorBodyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_bankease_bff-service_protogen_bff-service.InternetBillDetail": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "customerId": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "internetFee": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "tax": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "total": {
                    "type": "string"
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.InternetBillResponse": {
            "type": "object",
            "properties": {
                "bill": {
                    "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.InternetBillDetail"
                }
            }
        },
...
        "github_com_bankease_bff-service_protogen_bff-service.PaymentCardItem": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "balance": {
                    "type": "integer"
                },
                "brand": {
                    "type": "string"
                },
                "cardLabel": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "gradientColors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "holderName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "maskedNumber": {
                    "type": "string"
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.PaymentCardListResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.PaymentCardItem"
                    }
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.PrepaidPayRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "cardId": {
                    "type": "string"
                },
                "idempotencyKey": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.PrepaidPayResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_bankease_bff-service_protogen_bff-service.ProviderItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_bankease_bff-service_protogen_bff-service.ProviderListResponse": {
            "type": "object",
            "properties": {
                "providers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bankease_bff-service_protogen_bff-service.ProviderItem"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token with the `Bearer ` prefix, e.g. \"Bearer eyJhbGciOi...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}